#43 refactor: 이미지 업로드 유틸 리팩토링
