+++ CHICCHIC/src/apis/postApi.ts
@@ -1,10 +1,10 @@
-import { axiosInstance } from "./axios";
+import { axiosInstance } from "./axiosInstance";
 import type { PostCategory } from "../types/enums/postCategory";
 import type { ResponsePostListDto, RequestCreatePostDto } from "../types/post";
 import type { ResponseUploadImg } from "../types/img";
 
 // 향수 추천 상담소 게시글 리스트 조회
-export const getPostList = async (
+export const getConsultPostList = async (
   category: PostCategory
 ): Promise<ResponsePostListDto> => {
   const { data } = await axiosInstance.get("/consult-posts", {
@@ -22,12 +22,15 @@ export const getPostList = async (
 //   consultId,
 // }): Promise<>
 
-// 게시글 포스트에 대한 이미지 업로드
-export const uploadPostImg = async (file: File): Promise<ResponseUploadImg> => {
+// 이미지 업로드 유틸
+const upLoadImg = async (
+  file: File,
+  endpoint: string
+): Promise<ResponseUploadImg> => {
   try {
     const formData = new FormData();
     formData.append("file", file);
-    const { data } = await axiosInstance.post("/images/upload", formData);
+    const { data } = await axiosInstance.post(endpoint, formData);
     return data;
   } catch (e) {
     console.error("이미지 업로드 실패", e);
@@ -35,17 +38,11 @@ export const uploadPostImg = async (file: File): Promise<ResponseUploadImg> => {
   }
 };
 
+// 게시글 포스트에 대한 이미지 업로드
+export const uploadPostImg = (file: File) => {
+  upLoadImg(file, "/images/upload");
+};
 // 프로필 아바타 설정에 대한 이미지 업로드
-export const uploadProfileImg = async (
-  file: File
-): Promise<ResponseUploadImg> => {
-  try {
-    const formData = new FormData();
-    formData.append("file", file);
-    const { data } = await axiosInstance.post("/users/me", formData);
-    return data;
-  } catch (e) {
-    console.error("이미지 업로드 실패", e);
-    throw e;
-  }
+export const uploadProfileImg = (file: File) => {
+  upLoadImg(file, "/users/me");
 };
