+++ CHICCHIC/src/hooks/queries/usePostList.ts
@@ -2,12 +2,12 @@ import { useQuery } from "@tanstack/react-query";
 import type { PostCategory } from "../../types/enums/postCategory";
 import type { ResponsePostListDto } from "../../types/post";
 import { QUERY_KEY } from "../../constants/key";
-import { getPostList } from "../../apis/postApi";
+import { getConsultPostList } from "../../apis/postApi";
 
 export const usePostList = (category: PostCategory) => {
   return useQuery<ResponsePostListDto, Error>({
     queryKey: [QUERY_KEY.post, category],
-    queryFn: () => getPostList(category),
+    queryFn: () => getConsultPostList(category),
     // 3분간 캐시 유지
     staleTime: 1000 * 60 * 3,
   });
